Prep VS Code integration
